cloud-functions/dataforseo-bigquery-sync/requirements.txt
functions-framework==3.8.2
google-cloud-firestore==2.19.0
google-cloud-bigquery==3.28.0
